Binutils auto generated objdump improvements:
Introduce the concept of fields_positions, which contains an encoding
of the [start,end] bits of all fields in a operand in the same bitvector.

Also fix this *horrible* disasm code, which was wrong and not working.
Now, whenever we have an immediate, we print it as a signed number in
a comment after the instruction dump.