BUGFIX: Bug found when the memory device has two levels
The memory device has a number that shows to acsim your level of hierarchy. But if the model has only the data cache, the device memory will have two levels of hierarchy: level 0 to instruction flow and level 1 to data flow. So, the functions to find the first level data device and the fetch instruction device have been fixed.