[ac_syscall] Fix writev syscall to work on 64-bit hosts
Currently, all official ArchC models are 32-bit machines. This code
change forces writev to always read the guest parameters supposing
pointers are 4-byte wide. Previously, it assumed the guest had the
the same pointer size as the host, but this causes confusion because
ArchC syscalls worked on 32-bit machines but not on 64-bit machines.