Add get_buffer_addr() to the ac_syscall interface
Adds a new function to the syscall interface, which ArchC models need
to implement to have a working syscall functionality. This function
is a memcpy from guest to host memory with a slightly different
interface: the address of the memory location is given directly,
instead of indirectly via a guest register number.

This simplifies the implementation of the writev syscall and allows
MIPS syscalls to work.