Tentative solution to issue #45: https://github.com/ArchC/ArchC/issues/45 "id" register in the models must exist
I created a new ArchC register called ac_id and changed all generated code to use it instead of id. In this way, models do not need the id register and the processor id functionality continues to work.

The patch follows all ac_pc activity and create extra lines for ac_id. It also updates the interrupt handler to use ac_id instead of id.

For external use, the getId() method is still valid. So no change in non-ArchC code should be required.